# server/api/generated/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.11.0 DO NOT EDIT.
package openapi

import (
	"time"
)

// DefaultErrorResponse defines model for DefaultErrorResponse.
type DefaultErrorResponse struct {
	Message *string `json:"message,omitempty"`
}

// NewTodo defines model for NewTodo.
type NewTodo struct {
	Content *string `json:"content,omitempty"`
}

// Todo defines model for Todo.
type Todo struct {
	Content   *string    `json:"content,omitempty"`
	CreateAt  *time.Time `json:"createAt,omitempty"`
	Id        *int       `json:"id,omitempty"`
	UpdatedAt *time.Time `json:"updatedAt,omitempty"`
}

// UpdateTodoJSONBody defines parameters for UpdateTodo.
type UpdateTodoJSONBody = NewTodo

// PostTodoJSONBody defines parameters for PostTodo.
type PostTodoJSONBody = NewTodo

// UpdateTodoJSONRequestBody defines body for UpdateTodo for application/json ContentType.
type UpdateTodoJSONRequestBody = UpdateTodoJSONBody

// PostTodoJSONRequestBody defines body for PostTodo for application/json ContentType.
type PostTodoJSONRequestBody = PostTodoJSONBody